requirements.txt
fastapi==0.63.0
python-dotenv==0.14.0
sqlalchemy==1.4.3
psycopg2==2.8.6
uvicorn==0.13.4